Move StarDifficulty to own file
